requirements.txt
Django==5.1.2
pika==1.3.2
psycopg2-binary==2.9.10
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
model-bakery==1.20.0
celery==5.4.0